move: table component
- 렌더 디렉으로 옮김
